+++ CHANGELOG.md
@@ -20,6 +20,7 @@ We've also changed the following things in Server Manager itself:
 
 * Significantly improved startup times - we've optimised some of Server Manager's startup logic.
 * Added split information to the Current Lap column of the Live Timings page.
+* Added tyre information to Live Timings.
 
 Removed: